backend/requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.45.0
pandas==3.0.2
openpyxl==3.1.5
pyyaml==6.0.2
python-multipart==0.0.26
pytest==9.0.3
pytest-cov==7.1.0
